cleanup. grid parts
